+++ Entities/EnemyEntities/EnemyEntityWithProjectile.cs
@@ -17,7 +17,7 @@ namespace SprintZero1.Entities.EnemyEnetities
         /// <param name="position">The position of the enemy entity.</param>
         /// <param name="startingHealth">The starting health of the enemy entity.</param>
         /// <param name="enemyName">The name of the enemy.</param>
-        public EnemyEntityWithProjectile(Vector2 position, int startingHealth, string enemyName)
+        public EnemyEntityWithProjectile(Vector2 position, float startingHealth, string enemyName)
         : base(position, startingHealth, enemyName)
         {
             //no special constructor thing
